feat:added function to publish and share models

+++ app/Http/Controllers/ModelController.php
@@ -12,9 +12,14 @@ class ModelController extends Controller
     {
         $user = auth()->user();
         $user->load('models');
+
+        $followed_models = $user->followedModels()->get();
+        $followed_models->load('author');
+
         // inertia render
         return Inertia::render('Models/Index', [
             'user' => $user,
+            'followed_models' => $followed_models,
         ]);
     }
 
@@ -50,9 +55,16 @@ class ModelController extends Controller
     public function show(AIModel $model)
     {
         $user = auth()->user();
+        $model->load('author');
+        $is_author = $model->is_owner($user->id);
+        $is_followed = $user->followedModels()->where('model_id', $model->id)->count() > 0;
+
+
        return Inertia::render('Models/Show', [
             'model' => $model,
             'user' => $user,
+            'is_author' => $is_author,
+            'is_followed' => $is_followed,
         ]);
     }
 
@@ -80,11 +92,66 @@ class ModelController extends Controller
 
     public function destroy(AIModel $model)
     {
+
+        $user = auth()->user();
+        $author = $model->is_owner($user->id);
+        if(!$author){
+            return redirect()->route('models.index')->with('error', 'You are not the author of this model');
+        }
         $model->delete();
 
         return redirect()->route('models.index')
             ->with('success', 'Model deleted successfully');
     }
+
+
+    public function makePublic(AIModel $model)
+    {
+        $model->is_public = true;
+        $model->save();
+
+        $author = $model->is_owner(auth()->user()->id);
+        if(!$author){
+            return redirect()->route('models.index')->with('error', 'You are not the author of this model');
+        }
+
+
+        return redirect()->route('models.index')
+            ->with('success', 'Model made public successfully');
+    }
+
+    public function makePrivate(AIModel $model)
+    {
+        $model->is_public = false;
+        $model->save();
+
+        $author = $model->is_owner(auth()->user()->id);
+        if(!$author){
+            return redirect()->route('models.index')->with('error', 'You are not the author of this model');
+        }
+
+        return redirect()->route('models.index')
+            ->with('success', 'Model made private successfully');
+    }
+
+    public function follow(AIModel $model)
+    {
+        $user = auth()->user();
+        $user->followedModels()->attach($model->id);
+
+        return redirect()->route('models.index')
+            ->with('success', 'Model followed successfully');
+    }
+
+    public function unfollow(AIModel $model)
+    {
+        $user = auth()->user();
+        $user->followedModels()->detach($model->id);
+
+        return redirect()->route('models.index')
+            ->with('success', 'Model unfollowed successfully');
+    }
+
 
 
 }
